Move external service implementations to separate library.
