+++ src/AlphaID/AlphaID.sln
@@ -64,6 +64,10 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "tests", "tests", "{7714CF67
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "RealName", "RealName\RealName.csproj", "{F4F56EF1-A0A7-43C0-B65A-804DD84AF237}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AlphaIDPlatformServices", "AlphaIDPlatformServices\AlphaIDPlatformServices.csproj", "{023D269C-A84C-4E73-BEF1-5D721B73ADC9}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "XinAnAuth", "XinAnAuth\XinAnAuth.csproj", "{26C66A92-2A73-4A22-AED4-4BE92CA42AAF}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -150,6 +154,14 @@ Global
 		{F4F56EF1-A0A7-43C0-B65A-804DD84AF237}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{F4F56EF1-A0A7-43C0-B65A-804DD84AF237}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{F4F56EF1-A0A7-43C0-B65A-804DD84AF237}.Release|Any CPU.Build.0 = Release|Any CPU
+		{023D269C-A84C-4E73-BEF1-5D721B73ADC9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{023D269C-A84C-4E73-BEF1-5D721B73ADC9}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{023D269C-A84C-4E73-BEF1-5D721B73ADC9}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{023D269C-A84C-4E73-BEF1-5D721B73ADC9}.Release|Any CPU.Build.0 = Release|Any CPU
+		{26C66A92-2A73-4A22-AED4-4BE92CA42AAF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{26C66A92-2A73-4A22-AED4-4BE92CA42AAF}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{26C66A92-2A73-4A22-AED4-4BE92CA42AAF}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{26C66A92-2A73-4A22-AED4-4BE92CA42AAF}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -178,6 +190,8 @@ Global
 		{D2986112-01A1-445D-ADD2-399E9E483BD4} = {D92B699A-567D-4D77-B830-F9FE537BCC59}
 		{7714CF67-4098-4F37-92AD-C465912B0FB0} = {7CC468B3-52ED-4E35-BDB5-38642BE5B3C4}
 		{F4F56EF1-A0A7-43C0-B65A-804DD84AF237} = {D92B699A-567D-4D77-B830-F9FE537BCC59}
+		{023D269C-A84C-4E73-BEF1-5D721B73ADC9} = {8D38EFB9-849B-430A-8D9F-4D22524D658B}
+		{26C66A92-2A73-4A22-AED4-4BE92CA42AAF} = {8D38EFB9-849B-430A-8D9F-4D22524D658B}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {FAC996D9-8451-4B0E-99CC-F15843E0538A}
