+++ src/AlphaID/AlphaIDPlatformServices/Aliyun/AliyunChineseIDCardOCRService.cs
@@ -1,9 +1,10 @@
-﻿using Microsoft.Extensions.Options;
+﻿using AlphaIDPlatform.Platform;
+using Microsoft.Extensions.Options;
 using Newtonsoft.Json;
 using System.Net.Http.Headers;
 using System.Net.Http.Json;
 
-namespace AlphaIDPlatform.Platform;
+namespace AlphaIDPlatformServices.Aliyun;
 
 /// <summary>
 /// 阿里云身份证识别服务。
