+++ src/AlphaID/AlphaIDPlatformServices/Primitives/FreeTextMessage.cs
@@ -1,3 +1,3 @@
-﻿namespace AlphaIDPlatform.Platform;
+﻿namespace AlphaIDPlatformServices.Primitives;
 
 internal record FreeTextMessage(string[] Recipients, string Content);
