+++ src/AlphaID/AlphaIDPlatformServices/Primitives/SimpleShortMessageService.cs
@@ -1,9 +1,10 @@
-﻿using Microsoft.Extensions.Options;
+﻿using AlphaIDPlatform.Platform;
+using Microsoft.Extensions.Options;
 using Newtonsoft.Json;
 using System.Net.Http.Headers;
 using System.Net.Http.Json;
 
-namespace AlphaIDPlatform.Platform;
+namespace AlphaIDPlatformServices.Primitives;
 
 /// <summary>
 ///
