+++ src/AlphaID/AlphaIDPlatformServices/Primitives/SimpleShortMessageServiceOptions.cs
@@ -1,4 +1,4 @@
-﻿namespace AlphaIDPlatform.Platform;
+﻿namespace AlphaIDPlatformServices.Primitives;
 
 /// <summary>
 ///
