+++ src/AlphaID/WechatAuthWebApp/WechatAuthWebApp.csproj
@@ -17,6 +17,7 @@
 
 	<ItemGroup>
 		<ProjectReference Include="..\AlphaIDEntityFramework\AlphaIDEntityFramework.csproj" />
+		<ProjectReference Include="..\AlphaIDPlatformServices\AlphaIDPlatformServices.csproj" />
 		<ProjectReference Include="..\IDSubjects\IDSubjects.csproj" />
 		<ProjectReference Include="..\AlphaIDPlatform\AlphaIDPlatform.csproj" />
 		<ProjectReference Include="..\WechatWebLogin\WechatWebLogin.csproj" />
